requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning Libraries
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Streamlit and Web Framework
streamlit>=1.25.0
streamlit-aggrid>=0.3.4

# Data Processing and Visualization
scipy>=1.11.0
statsmodels>=0.14.0

# Database and Data Storage
sqlalchemy>=2.0.0
# sqlite3
pymongo>=4.4.0
redis>=4.6.0

# Utilities and Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
joblib>=1.3.0

# API and Web Services
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Jupyter and Development
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Optional: Advanced Analytics
networkx>=3.1.0
nltk>=3.8.0
textblob>=0.17.0

# Optional: Cloud and Deployment
boto3>=1.28.0
docker>=6.1.0
gunicorn>=21.2.0

# Optional: Additional Visualization
bokeh>=3.2.0
altair>=5.0.0

# Optional: Time Series
prophet>=1.1.0
pmdarima>=2.0.0

# Optional: Deep Learning (if needed)
# tensorflow>=2.13.0
# torch>=2.0.0
# transformers>=4.30.0

# Version constraints for compatibility
python-dateutil>=2.8.2
pytz>=2023.3
six>=1.16.0

# Development and Environment
pathlib2>=2.3.7
tqdm>=4.65.0
click>=8.1.0

# Data validation
pydantic>=2.0.0
marshmallow>=3.20.0

# Caching and Performance  
diskcache>=5.6.0
memory-profiler>=0.61.0

# Optional: Geospatial Analysis
# folium>=0.14.0
# geopandas>=0.13.0

# Optional: Financial Analysis
# yfinance>=0.2.0
# quantlib>=1.31.0